moves hnc_segmentation_utils into detect/
